Changes requested in PR#68. BatchWriter code updated and moved to misc_utils, tidy up query_mag,
