+++ nesta/production/routines/arxiv/arxiv_mag_task.py
@@ -11,7 +11,8 @@ import logging
 import pprint
 
 from arxiv_iterative_date_task import DateTask
-from nesta.packages.arxiv.collect_arxiv import BatchWriter, BatchedTitles, update_existing_articles
+from nesta.packages.arxiv.collect_arxiv import BatchedTitles, update_existing_articles
+from nesta.packages.misc_utils.batches import BatchWriter
 from nesta.packages.mag.query_mag import build_expr, query_mag_api, dedupe_entities, update_field_of_study_ids
 from nesta.production.orms.arxiv_orm import Base, Article
 from nesta.production.orms.mag_orm import FieldOfStudy
